arch: split index-access large-union/intersection helpers into sibling module
The intersection deferral guard pushed evaluate_rules/index_access.rs to
1856 lines, over its 1835 arch size ratchet (merge-queue arch-size
ejection). Move the literal-key large-union fast path and the
intersection receiver property-collection/this-property helpers to
index_access_fastpath_intersection.rs following the existing sibling
pattern; index_access.rs is now 1636 lines. No behavior change.
