+++ crates/tsz-solver/src/evaluation/evaluate_rules/mod.rs
@@ -20,6 +20,7 @@ mod index_access_augmented_redirect;
 mod index_access_callable;
 mod index_access_empty_key;
 mod index_access_evaluator_helpers;
+mod index_access_fastpath_intersection;
 mod index_access_keys;
 mod index_access_object_with_index;
 mod index_access_optional_read;
